requirements.txt
djangorestframework
django